pyproject.toml
[project]
name            = "scratchimg"
version         = "0.1.0"
description     = "Add your description here"
authors         = [{ name = "aspizu", email = "aspizu@protonmail.com" }]
dependencies    = ["pillow>=10.3.0", "msgspec>=0.18.6"]
readme          = "README.md"
requires-python = ">= 3.8"

[build-system]
requires      = ["hatchling"]
build-backend = "hatchling.build"

[tool.rye]
managed          = true
dev-dependencies = []

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/scratchimg"]

[tool.rye.scripts]
main = { call = "src.scratchimg:main" }

[tool.ruff.lint]
select = ["ALL"]
extend-safe-fixes = ["ALL"]
extend-unsafe-fixes = ["T201"]
unfixable = []
ignore = [
    "W191",
    "E111",
    "E114",
    "E117",
    "D",
    "Q000",
    "Q001",
    "Q002",
    "Q003",
    "COM812",
    "COM819",
    "ISC001",
    "ISC002",
    "ANN101",
    "ANN102",
    "SLF001",
    "FBT",
    "TD002",
    "TD003",
    "E501",
    'PLW2901',
    'PGH003',
]

[tool.pyright]
typeCheckingMode                   = "strict"
reportUnnecessaryTypeIgnoreComment = true
reportUnknownArgumentType          = false
reportUnknownLambdaType            = false
reportUnknownMemberType            = false
reportUnknownParameterType         = false
reportUnknownVariableType          = false

[tool.ruff.lint.isort]
required-imports = ["from __future__ import annotations"]

[tool.ruff.format]
docstring-code-format = true
